refactor(api): shared util
