+++ api/src/utils/string.ts
@@ -89,3 +89,35 @@ export const cleanIdParam = (param: string): string => {
 };
 
 export const isBlank = (value?: string | null) => value === null || value === undefined || value === "";
+
+/**
+ * Jaccard similarity between two underscore-separated token strings.
+ * Order-insensitive: "sante_social" and "social_sante" → 1.0
+ */
+export const jaccardSimilarity = (a: string, b: string): number => {
+  const setA = new Set(a.split("_"));
+  const setB = new Set(b.split("_"));
+  const intersection = new Set([...setA].filter((t) => setB.has(t)));
+  const union = new Set([...setA, ...setB]);
+  return intersection.size / union.size;
+};
+
+/**
+ * Find the closest match for a candidate string among a set of valid keys,
+ * using Jaccard similarity on underscore-split tokens.
+ * Returns the best match if its score >= threshold, null otherwise.
+ */
+export const fuzzyMatchKey = (
+  candidate: string,
+  validKeys: Iterable<string>,
+  threshold: number
+): { key: string; score: number } | null => {
+  let best: { key: string; score: number } | null = null;
+  for (const key of validKeys) {
+    const score = jaccardSimilarity(candidate, key);
+    if (score >= threshold && (!best || score > best.score)) {
+      best = { key, score };
+    }
+  }
+  return best;
+};
